Move things from ArborX_DetailsAlgorithms.hpp to geometry files
