+++ src/geometry/ArborX_Box.hpp
@@ -15,6 +15,7 @@
 #include <ArborX_DetailsKokkosExtArithmeticTraits.hpp>
 #include <ArborX_DetailsKokkosExtMinMaxOperations.hpp>
 #include <ArborX_GeometryTraits.hpp>
+#include <ArborX_HyperBox.hpp> // for algorithms
 #include <ArborX_Point.hpp>
 
 #include <Kokkos_Macros.hpp>
