+++ src/geometry/ArborX_KDOP.hpp
@@ -337,4 +337,24 @@ struct ArborX::GeometryTraits::coordinate_type<ArborX::Experimental::KDOP<k>>
   using type = float;
 };
 
+namespace ArborX::Details::Dispatch
+{
+using GeometryTraits::BoxTag;
+using GeometryTraits::KDOPTag;
+
+// expand a box to include a kdop
+template <typename Box, typename KDOP>
+struct expand<BoxTag, KDOPTag, Box, KDOP>
+{
+  KOKKOS_FUNCTION static void apply(Box &box, KDOP const &kdop)
+  {
+    // FIXME This is a workaround so that we can use existing conversion
+    // machinery for KDOP. In the long term, this should be replaced by a
+    // general algorithm.
+    Details::expand(box, (ArborX::Box)kdop);
+  }
+};
+
+} // namespace ArborX::Details::Dispatch
+
 #endif
